fix browser tests
