+++ tests/Browser/Pages/WalletTest.php
@@ -1328,7 +1328,7 @@ describe('Tokens Tab', function () {
                     ->waitForText('5 results', ignoreCase: true);
 
                 foreach ($tokenHolders as $tokenHolder) {
-                    $browser->assertSee($tokenHolder->token->name)
+                    $browser->assertSee($tokenHolder->token->nameNormalized)
                         ->assertSee($tokenHolder->token->symbol)
                         ->assertSee(substr($tokenHolder->token_address, 0, 5).'…'.substr($tokenHolder->token_address, -5))
                         ->assertSee(number_format($tokenHolder->balance->toFloat(), 4));
@@ -1351,8 +1351,8 @@ describe('Tokens Tab', function () {
                     ->pause(100)
                     ->waitForText('1 result', ignoreCase: true);
 
-                $browser->assertSee($tokenHolder->token->name)
-                    ->assertSee($tokenHolder->token->symbol)
+                $browser->assertSee($tokenHolder->token->nameNormalized)
+                    ->assertSee($tokenHolder->token->symbolNormalized)
                     ->assertSee(substr($tokenHolder->token_address, 0, 5).'…'.substr($tokenHolder->token_address, -5));
 
                 $selector = '[data-testid="token:'.$tokenHolder->token->symbol.':amount"]';
@@ -1390,13 +1390,13 @@ describe('Tokens Tab', function () {
 
             $browser->visitRoute('wallet', ['wallet' => $this->wallet, 'view' => 'tokens'])
                 ->waitForText('50 results', ignoreCase: true)
-                ->assertSee($sortedTokens->first()->token->name)
-                ->assertSee($sortedTokens->take(25)->last()->token->name)
+                ->assertSee($sortedTokens->first()->token->nameNormalized)
+                ->assertSee($sortedTokens->take(25)->last()->token->nameNormalized)
                 ->click('[data-testid="pagination:next-page"] button')
                 ->waitForText('Page 2 of 2')
                 ->assertQueryStringHas('page', '2')
-                ->assertSee($sortedTokens->skip(25)->first()->token->name)
-                ->assertSee($sortedTokens->skip(25)->take(25)->last()->token->name);
+                ->assertSee($sortedTokens->skip(25)->first()->token->nameNormalized)
+                ->assertSee($sortedTokens->skip(25)->take(25)->last()->token->nameNormalized);
         });
     })->with('desktop_mobile_resolutions');
 
@@ -1416,15 +1416,15 @@ describe('Tokens Tab', function () {
             $browser->visitRoute('wallet', ['wallet' => $this->wallet, 'view' => 'tokens', 'page' => 2])
                 ->waitForText('50 results', ignoreCase: true)
                 ->assertSee('Page 2 of 2')
-                ->assertDontSee($sortedTokens->first()->token->name)
-                ->assertDontSee($sortedTokens->take(25)->last()->token->name)
+                ->assertDontSee($sortedTokens->first()->token->nameNormalized)
+                ->assertDontSee($sortedTokens->take(25)->last()->token->nameNormalized)
                 ->click('[data-testid="pagination:per-page-dropdown:button"]')
                 ->waitForTextIn('[data-testid="pagination:per-page-dropdown:dropdown"]', '10')
                 ->clickAtXPath('//div[@data-testid="pagination:per-page-dropdown:dropdown"]//div[normalize-space(text())="10"]')
                 ->waitForText('Page 1 of 5')
-                ->assertSee($sortedTokens->first()->token->name)
-                ->assertSee($sortedTokens->take(10)->last()->token->name)
-                ->assertDontSee($sortedTokens->take(11)->last()->token->name);
+                ->assertSee($sortedTokens->first()->token->nameNormalized)
+                ->assertSee($sortedTokens->take(10)->last()->token->nameNormalized)
+                ->assertDontSee($sortedTokens->take(11)->last()->token->nameNormalized);
         });
     })->with('desktop_mobile_resolutions');
 });
